+++ CHANGES.rst
@@ -4,6 +4,7 @@ Changes
 Version 0.9.2 (UNRELEASED)
 --------------------------
 
+- Adds option to delete all the runs of a workflow.
 - Changes Docker image Node version from 16 to 18.
 
 Version 0.9.1 (2023-09-27)